Adds scaffolding code and some basic functionality to start with.
Adds scaffolding setup with npm, webpack, babel, bootstrap, etc
Seems to work, not thoroughly tested.

Adds some functionality to start with. Code is not super clean.

Trying to work fast due to time constraint so I'm not writing
as clean commits as usually. Rebasing later.

Perhaps I'll allow console.log() and not remove it when rebasing.
It's a matter of taste how clean the code should look when in progress.
Possible to squash a lot and thereby get rid of the console.log().

Note that API key value is stored in an .env file so not visible here.